+++ src/components/CheckboxGroup/CheckboxGroup.test.tsx
@@ -1,6 +1,7 @@
 import { Flex, Spacer } from "@chakra-ui/react";
-import * as React from "react";
 import { render, screen } from "@testing-library/react";
+import userEvent from "@testing-library/user-event";
+import * as React from "react";
 import { axe } from "jest-axe";
 import renderer from "react-test-renderer";
 
@@ -8,16 +9,49 @@ import * as generateUUID from "../../helpers/generateUUID";
 import CheckboxGroup from "./CheckboxGroup";
 import Checkbox from "../Checkbox/Checkbox";
 import { CheckboxGroupLayoutTypes } from "./CheckboxGroupLayoutTypes";
-import userEvent from "@testing-library/user-event";
 
-describe("Checkbox Accessibility", () => {
-  it("passes axe accessibility", async () => {
+describe("CheckboxGroup Accessibility", () => {
+  it("passes axe accessibility with string labels ", async () => {
     const { container } = render(
       <CheckboxGroup labelText="CheckboxGroup example" name="a11y-test">
         <Checkbox value="2" labelText="Checkbox 2" />
         <Checkbox value="3" labelText="Checkbox 3" />
         <Checkbox value="4" labelText="Checkbox 4" />
         <Checkbox value="5" labelText="Checkbox 5" />
+      </CheckboxGroup>
+    );
+    expect(await axe(container)).toHaveNoViolations();
+  });
+  it("passes axe accessibility with jsx labels", async () => {
+    const { container } = render(
+      <CheckboxGroup
+        labelText="jsxLabels"
+        name="jsxLabels"
+        id="jsxLabels"
+        isFullWidth
+      >
+        <Checkbox
+          id="arts"
+          labelText={
+            <Flex>
+              <span>Arts</span>
+              <Spacer />
+              <span>4</span>
+            </Flex>
+          }
+          value="arts"
+        />
+        <Checkbox
+          id="English"
+          labelText={
+            <Flex>
+              <span>English</span>
+              <Spacer />
+              <span>23</span>
+            </Flex>
+          }
+          value="English"
+        />
       </CheckboxGroup>
     );
     expect(await axe(container)).toHaveNoViolations();
@@ -120,8 +154,8 @@ describe("Checkbox", () => {
   });
 
   it("sets the next value through the onChange function", () => {
-    let newValue = [];
-    const onChange = (value) => {
+    let newValue: string[] = [];
+    const onChange = (value: string[]) => {
       newValue = value;
     };
     render(
@@ -350,10 +384,10 @@ describe("Checkbox", () => {
     const withJSXCheckboxLabels = renderer
       .create(
         <CheckboxGroup
-          labelText="jsxLabels"
-          name="jsxLabels"
           id="jsxLabels"
           isFullWidth
+          labelText="jsxLabels"
+          name="jsxLabels"
         >
           <Checkbox
             id="arts"