+++ src/theme/components/checkbox.ts
@@ -69,10 +69,7 @@ const baseStyleControl = {
 };
 
 // Style object for the Checkbox's label
-const baseStyleLabel = {
-  ...checkboxRadioLabelStyles,
-  width: "100%",
-};
+const baseStyleLabel = checkboxRadioLabelStyles;
 
 // Style object for the Checkbox's helper text
 const baseStyleHelper = checkboxRadioHelperStyle;