+++ src/theme/components/checkboxGroup.ts
@@ -3,10 +3,7 @@ import { checkboxRadioGroupStyles } from "./global";
 const CheckboxGroup = {
   parts: ["helper", "stack"],
   baseStyle: ({ isFullWidth = false }) => ({
-    ...checkboxRadioGroupStyles,
-    stack: {
-      width: isFullWidth ? "100%" : checkboxRadioGroupStyles.stack.width,
-    },
+    ...checkboxRadioGroupStyles(isFullWidth),
   }),
 };